Implemented Ego/Allo Q Learning. Removed Deadline as a parameter. Answered states question more thoroughly
Implementing Ego/Allo Q learning as mentioned in the Studywolf blog
(also addressed in my report). Did not see much or any improvements in
cumulative rewards over all the trials. Need to check out other metrics.

Realized Deadline parameter in the State decreased the Agents ability
to learn quickly. This is due to the deadline increasing the dimensions
of the state space 30 or more times. Removing deadline allows the agent
to find the deadline almost every time.