Cleaned up analysis. Discretized deadline. Updated Report.
Consolidated all for loops over the output analysis to one for loop and
added percent of trials completed to analysis.

Made a more discrete deadline that consolidates the deadline values
into fractions of the deadline value. Specifically, thirds, then
halves. Neither proved to be much use.

Updated the report with my findings.